requirements.txt
langchain==0.2.10
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-community==0.2.10
streamlit